% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_addimages.R
\name{pie_addimages}
\alias{pie_addimages}
\title{Add images to pie charts}
\usage{
pie_addimages(
  mypie,
  imglist,
  perc = "no",
  lbl_col = "black",
  border_col = "black"
)
}
\arguments{
\item{mypie}{A ggplot object. The output from \code{\link{pie_bake}} function,
\strong{NOT from \code{\link{pie_bake_pro}}}. \cr
Don't focus on the template, only lables and values will be inherited.}

\item{imglist}{A list of objects returned by and \code{\link[jpeg]{readJPEG}}
used to fill slices.}

\item{perc}{You can choose among:\cr
\itemize{
\item 'no'    : the percentages won't be displayed (default choice);
\item 'below' : the percentages will be displayed below the group labels;
\item 'right' : the percentages will be displayed next to the group labels.
}}

\item{lbl_col}{A string containing the chosen color for the labels, default is
\code{"black"}.}

\item{border_col}{A string containing the chosen color for the border of the pie
chart, default is \code{"black"}.}
}
\value{
A ggplot object.
}
\description{
This function allows to easily add some figures to you pie chart, taking information
from the output of the \code{\link{pie_bake}} function.
You can also customize the chart by specifying the colors of borders and labels
and where (and if) you want to print the percentages for each group.
}
\examples{
img1 <- jpeg::readJPEG(system.file("img", "mark.jpeg", package="tastypie"))
img2 <- jpeg::readJPEG(system.file("img", "brownie.jpeg", package="tastypie"))
img3 <- jpeg::readJPEG(system.file("img", "logo.jpeg", package="tastypie"))
imgs <- list(img1,img2,img3)

 df <- data.frame(
   c("Mark fan", "My brownie", "Tastypie logo"),
   c(500, 200, 300)
 )

 mypie <- pie_bake(df, template = "basic1")

 pie_addimages(
   mypie = mypie,
   imglist = imgs,
   perc = "below",
   lbl_col = "red",
   border_col = "grey"
 )

}
