% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_bake_pro.R
\name{pie_bake_pro}
\alias{pie_bake_pro}
\title{Easily create (more complex) pie charts}
\usage{
pie_bake_pro(data, template, group_name = "group", title = "")
}
\arguments{
\item{data}{A data frame with two variables (columns):\cr
\itemize{
\item in the first one there must be the vector of labels;\cr
\item in the second one there must be the vector of values.
}

You can use \code{\link{pie_datacheck}} to understand if the data is suitable.}

\item{template}{The chosen template.\cr
Type \code{pie_template_list_pro} to display all the available ones for this function.}

\item{group_name}{A string. If you want, you can specify a name for the categories.}

\item{title}{A string. If you want, you can specify the title of the graph.}
}
\description{
This function allows you to create cool pie charts easily by providing just
the data and a template among the available ones. Moreover, you can optionally
specify a title for the chart and a name for the categories.\cr \cr
Using this function you can create some \emph{complex and extravagant} pie charts.
If you are looking for something more \emph{classical} (and probably more understandable),
check out the \code{\link{pie_bake}} function.
}
\examples{
example <- data.frame(
  c("a. This", "b. Is", "c. Just", "d. An", "e. Example"),
  c(2.9, 6.9, 4.20, 13.12, 6.66)
)
pie_bake_pro(
  data = example,
  template = "eaten3",
  group_name = "cat:",
  title = "Example1"
)

pie_bake_pro(
  data = example,
  template = "dart1",
  title = "Example2!"
)

pie_bake_pro(
  data = example,
  template = "eye5",
  group_name = "GROUPS:"
)


}
