% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pie_template_list}
\alias{pie_template_list}
\title{Available templates in \code{tastypie}.}
\format{
An object of class \code{character} of length 30.
}
\usage{
pie_template_list
}
\description{
The vector containing all the available templates for pie charts for the \code{tastypie}
package.
}
\keyword{datasets}
