% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2.R
\docType{data}
\name{pie_template_list_pro}
\alias{pie_template_list_pro}
\title{Available templates in \code{tastypie} for \code{pie_bake_pro}.}
\format{
An object of class \code{character} of length 30.
}
\usage{
pie_template_list_pro
}
\description{
The vector containing all the available templates for pie charts for the
\code{\link{pie_bake_pro}} function.
Other less complex templates are available in \code{\link{pie_template_list}}
that can be used through the \code{\link{pie_bake}} function.
}
\seealso{
See all the available templates displayed
\href{https://paolodalena.github.io/tastypie/articles/available_templates.html}{here}!
}
\keyword{datasets}
