% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tastypie.R
\docType{package}
\name{tastypie}
\alias{tastypie}
\title{Easy pie charts with \code{tastypie}}
\description{
\figure{logo.png}{options: alt = 'logo' width = 80}

You only need to type
\href{https://www.google.com/search?q=why+pie+charts+are+bad}{why pie charts are bad}
on Google to find thousands of articles full of (\emph{valid}) reasons why other types of charts
should be preferred over this one.\cr
Therefore, because of the little use due to the reasons already mentioned,
making pie charts (and related) in R is not straightforward, so other functions are needed
to simplify things.\cr
In this R package there are useful functions to make \emph{tasty} pie charts immediately
by exploiting the many cool templates provided.\cr\cr
}
\seealso{
Find more information on the package web site
\href{https://paolodalena.github.io/tastypie/}{here}, or on the GitHub package page
\href{https://github.com/PaoloDalena/tastypie}{here}.
}
\author{
\strong{Author and Maintainter}: Paolo Dalena \email{paolodalena97@gmail.com}
}
